CP5 User Roles

+++ app/controllers/wikis_controller.rb
@@ -1,57 +1,66 @@
+
 class WikisController < ApplicationController
   def index
-  	@wikis = Wiki.all
+    @wikis = Wiki.all
   end
 
   def show
-  	@wiki = Wiki.find(params[:id])
+    @wiki = Wiki.find(params[:id])
   end
 
   def new
-  	@wiki = Wiki.new
+    @wiki = Wiki.new
+    authorize @wiki
+  end
+
+  def edit
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def create
     @wiki = Wiki.new
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
-  
+    @wiki.private = params[:wiki][:private]
+    @wiki.user = current_user
+    authorize @wiki
+    
     if @wiki.save
-      flash[:notice] = "Wiki was saved."
+      flash[:notice] = "Your wiki was saved successfully."
       redirect_to @wiki
     else
-      flash.now[:alert] = "There was an error saving the wiki. Please try again."
+      flash.now[:alert] = "There was an error saving your wiki. Please try again later."
       render :new
     end
-  end
-
-  def edit
-  	@wiki = Wiki.find(params[:id])
   end
 
   def update
     @wiki = Wiki.find(params[:id])
     @wiki.title = params[:wiki][:title]
     @wiki.body = params[:wiki][:body]
-
+    @wiki.private = params[:wiki][:private]
+    authorize @wiki
+    
     if @wiki.save
-      flash[:notice] = "Wiki was saved."
+      flash[:notice] = "Wiki was updated successfully."
       redirect_to @wiki
     else
       flash.now[:alert] = "There was an error saving the wiki. Please try again."
-      render :new
+      render :edit
     end
   end
 
-	def destroy
-	  @wiki = Wiki.find(params[:id])
-
-	  if @wiki.destroy
-	    flash[:notice] = "\"#{@wiki.title}\" has been deleted."
-	    redirect_to wikis_path
-	  else
-	    flash.now[:alert] = "There was an error deleting the wiki. Please try again."
-	    render :show
-	  end
-	end
+  def destroy
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
+    
+    if @wiki.destroy
+      flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
+      redirect_to wikis_path
+    else
+      flash.now[:alert] = "There was an error deleting the wiki."
+      render :show
+    end
+  end
 end
